% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_utils.R
\name{add_caption}
\alias{add_caption}
\title{add_caption}
\usage{
add_caption(
  widget,
  caption,
  font_size = 12,
  align = "center",
  colour = "#555",
  italic = FALSE
)
}
\arguments{
\item{widget}{A networkD3::sankeyNetwork() htmlwidget.}

\item{caption}{Character caption text to place below the widget.}

\item{font_size}{Numeric font size in pixels.}

\item{align}{Character alignment: "left", "center", or "right".}

\item{colour}{Character CSS colour for the caption text.}

\item{italic}{Logical; if TRUE, render caption in italics.}
}
\description{
add_caption
}
