% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_utils.R
\name{add_linebreaks_in_labels}
\alias{add_linebreaks_in_labels}
\title{add_linebreaks_in_labels}
\usage{
add_linebreaks_in_labels(widget)
}
\arguments{
\item{widget}{A networkD3::sankeyNetwork() htmlwidget.}
}
\description{
add_linebreaks_in_labels
}
