% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_utils.R
\name{add_link_tooltips}
\alias{add_link_tooltips}
\title{add_link_tooltips}
\usage{
add_link_tooltips(widget, tooltips, ensure_hover = TRUE)
}
\arguments{
\item{widget}{A networkD3::sankeyNetwork() htmlwidget.}

\item{tooltips}{Character vector of tooltips, one per link (order must match Links).}

\item{ensure_hover}{Logical; if TRUE, enable pointer events on links.}
}
\description{
add_link_tooltips
}
