% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_utils.R
\name{add_node_tooltips}
\alias{add_node_tooltips}
\title{add_node_tooltips}
\usage{
add_node_tooltips(widget, tooltip)
}
\arguments{
\item{widget}{A networkD3::sankeyNetwork() htmlwidget.}

\item{tooltip}{Character vector of tooltips, one per node (order must match Nodes).}
}
\description{
add_node_tooltips
}
