% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_utils.R
\name{add_title}
\alias{add_title}
\title{add_title}
\usage{
add_title(widget, title, font_size = 18, align = "center")
}
\arguments{
\item{widget}{A networkD3::sankeyNetwork() htmlwidget.}

\item{title}{Character title text to place above the widget.}

\item{font_size}{Numeric font size in pixels.}

\item{align}{Character alignment: "left", "center", or "right".}
}
\description{
add_title
}
