% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{apply_filters}
\alias{apply_filters}
\title{Apply Filter Selections}
\usage{
apply_filters(data, filter_vars, input)
}
\arguments{
\item{data}{A data frame to be filtered}

\item{filter_vars}{Character vector of column names to filter on}

\item{input}{Shiny input object containing filter selections}
}
\value{
A filtered data frame
}
\description{
Subsets the data based on user filter selections from Shiny inputs.
Filters are only applied when selection is not NULL or "All".
}
