% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_utils.R
\name{center_node_labels}
\alias{center_node_labels}
\title{center_node_labels}
\usage{
center_node_labels(widget, nudge = 0)
}
\arguments{
\item{widget}{A networkD3::sankeyNetwork() htmlwidget.}

\item{nudge}{Numeric horizontal pixel offset to fine-tune centring.}
}
\description{
center_node_labels
}
