% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_labels}
\alias{flow_labels}
\title{Package flow labels}
\format{
A data frame with 39 rows and 3 variables:
\describe{
\item{flow_start}{Character string describing the intervention package (e.g. \code{"cm"}, \code{"cm, itn"}) of flow origin.}
\item{flow_end}{Character string describing the intervention package (e.g. \code{"cm"}, \code{"cm, itn"}) of flow terminus}
\item{flow_label}{Character string of label to display for flow.}
}
}
\source{
M3CPI project dataset
}
\usage{
flow_labels
}
\description{
A dataset with flow start and end codes and associated user-facing flow label
}
\keyword{datasets}
