% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{get_filter_vars}
\alias{get_filter_vars}
\title{Get Filter Variables}
\usage{
get_filter_vars(data)
}
\arguments{
\item{data}{A data frame containing the pathway data}
}
\value{
A character vector of column names to be used as filter variables
}
\description{
Identifies which columns in the data should be used as filter variables
by excluding core Sankey construction columns.
}
