% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_utils.R
\name{make_colour_scale}
\alias{make_colour_scale}
\title{make_colour_scale}
\usage{
make_colour_scale(unknown = "#cccccc")
}
\arguments{
\item{unknown}{Character CSS colour used for unmapped/unknown groups.}
}
\description{
make_colour_scale
}
