% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{make_nodes}
\alias{make_nodes}
\title{Create nodes for a step-package Sankey diagram}
\usage{
make_nodes(x)
}
\arguments{
\item{x}{A data frame containing at least a \code{package} column where each
entry describes the package combination at a given step.}
}
\value{
A data frame with columns \code{name}, \code{node_name}, and \code{id} suitable for
use with \code{\link[networkD3:sankeyNetwork]{networkD3::sankeyNetwork()}}.
}
\description{
Create nodes for a step-package Sankey diagram
}
\examples{
df <- simulate(3)
make_nodes(df)

}
