% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{make_sankey}
\alias{make_sankey}
\title{Render a step-package Sankey diagram}
\usage{
make_sankey(nodes, links, ...)
}
\arguments{
\item{nodes}{Data frame describing Sankey nodes as produced by
\code{\link[=make_nodes]{make_nodes()}}.}

\item{links}{Data frame describing connections between nodes, typically the
output of \code{\link[=make_links]{make_links()}}.}

\item{...}{Addtional arguments to pass to \code{\link[networkD3:sankeyNetwork]{networkD3::sankeyNetwork()}}}
}
\value{
An htmlwidgets object produced by
\code{\link[networkD3:sankeyNetwork]{networkD3::sankeyNetwork()}}.
}
\description{
Render a step-package Sankey diagram
}
