% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_steps.R
\name{simulate_step_data}
\alias{simulate_step_data}
\title{Simulate prioritisation step data}
\usage{
simulate_step_data(
  n = 10,
  all_interventions = c("ITN", "Chemoprevention", "Vaccine", "CM"),
  removable = c("ITN", "Chemoprevention", "Vaccine"),
  pfpr = c("low", "high"),
  seasonality = c("low", "high")
)
}
\arguments{
\item{n}{Number of simulations per grouping combination.}

\item{all_interventions}{All possible interventions.}

\item{removable}{Interventions that can be added or removed.}

\item{pfpr}{Levels of PfPr grouping variable.}

\item{seasonality}{Levels of seasonality grouping variable.}
}
\value{
A data frame with columns for grouping variables, simulation id,
step number, and package label.
}
\description{
Generate example step and package data for testing Sankey diagrams.
}
