% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_utils.R
\name{add_link_tooltips}
\alias{add_link_tooltips}
\title{add_link_tooltips}
\usage{
add_link_tooltips(widget, tooltip)
}
\arguments{
\item{widget}{A networkD3::sankeyNetwork() htmlwidget.}

\item{tooltip}{Character vector of tooltips, one per link (order must match Links).}
}
\description{
add_link_tooltips
}
